% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_text.R
\name{join_text}
\alias{join_text}
\title{Join delimited text file.}
\usage{
join_text(path = ".", pattern = ".csv", header = FALSE, sep = ";",
  dec = ",", quote = "", encoding = "UTF-8")
}
\arguments{
\item{path}{String. A path file.}

\item{pattern}{String. A pattern file like .txt or .csv.}

\item{header}{Logical. TRUE if the file has a header collum.}

\item{sep}{String. Set the separator character. ';' is default.}

\item{dec}{String. Set the decimal indicator. ',' is default.}

\item{quote}{Strig. the set of quoting characters. "\"" is dafault.}

\item{enconding}{String. Set the file enconding. 'UTF-8' is default.}
}
\value{
A unique file from all files in join.
}
\description{
The package pass several commum arguments like read_csv for join multiples
text files. You need pass a folder path. The function enters the specified
folder and joins all files and returns an object in your R. workspace.
}
\examples{
\dontrun{
# Call the function with the path is needed.
#join_text(path='~/folder/, pattern='.txt', sep=',')
}
}
