% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{Load frequent used packages.}
\usage{
load_pkg(x = c(), y = NULL)
}
\arguments{
\item{x}{A string vector of packages names (optional).}

\item{y}{Logical.}
}
\value{
Packages loaded.
}
\description{
If \code(x) do not specified, the package load a list of default packages.
}
\examples{

}
