% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhas_cores.R
\name{minhas_cores}
\alias{minhas_cores}
\title{A Set of a custom color pallete}
\usage{
minhas_cores(cor, indice)
}
\arguments{
\item{cor}{String A color name}

\item{indice}{Numéric (index) that indicate the color tones. 1 to 5.}
}
\description{
The function call an color hex from a intern list. It has five different
tones for each color that you can control by the index parameter.
}
