% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_flat.R
\name{theme_flat}
\alias{theme_flat}
\title{More one ggplot theme and colours}
\usage{
theme_flat(angle = NULL, ...)
}
\arguments{
\item{angle}{Optional. Change the default angle of axis.x text}
}
\description{
A custom flat ggplot theme and colorfull pallete.
}
