% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_text.R
\name{join_text}
\alias{join_text}
\title{Load frequent used packages.}
\usage{
join_text(path, pattern = ".csv", header = FALSE, sep = ";",
  dec = ",", enconding = "UTF8")
}
\arguments{
\item{path}{String. A path file.}

\item{pattern}{String. A pattern file like .txt or .csv.}

\item{header}{Logical. TRUE if the file has a header collum.}
}
\value{
A unique file from all files in join.
}
\description{
The package pass two argments x and y. x is a vector with packages that user plan to load. y is a logical vector, where TRUE load a default package list from developer.
You need chosse Y TRUE or FALSE. If x do not specified, the package load a list of default packages.
}
\examples{
join_text(path='./data/', pattern='.csv', header = T)


}
